refactor: refactor formula and zoom controller
